Added a utility getter function (to be used later on) to both Landmark and Station data models.

+++ app/src/main/java/co/swatisi/team/metrowalkdc/model/Landmark.kt
@@ -3,4 +3,4 @@ package co.swatisi.team.metrowalkdc.model
 import android.net.Uri
 
 data class Landmark(val name: String, val imageURL: Uri, val displayAddress: String,
-                    val lat: Float, val lon: Float, val distance: Float)
+                    val lat: Double, val lon: Double, val distance: Float)
